+++ mal_rustc_core/src/lib.rs
@@ -1,2 +1,4 @@
 pub mod parser;
 pub mod types;
+
+pub static MAL_RUNTIME: &'static str = include_str!("types/runtime.rs");
